Implemented a basic colour scheme again in the Density visualization.
Unlike the original in the QGraphicsScene version, the colours are actually
derived from the data the associated rectangles represent. qHash() is called
upon the address and the allocation time of the block, and from there, a
colour is found.

To be exact, only the first six bits are actually used. The six bits are
divided into three bits each (using the masks 0x07, 0x38, and 0x1c0), bit-
shifted as required to provide a red, green, or blue value, respectively.

Now: the next feature to work on is the scope resolution and transfer. That
means digging into the monitor again, but oh well. There will be one
modification to the scope resolution system, however . . .

Rather than grabbing the value of the return pointer in the overloaded malloc,
instead, the full backtrace will be produced. This is accomplished fairly
easily, as the old base pointer is stored in a position in memory that is
easy to retreieve. The difficulty is that I need to read up on how to access
C-style variables from assembly when said assembly is mixed into said C
function . . .