Fixed a few memory errors.
Nothing too special, only that I forgot to implement support for freeing the
data element of a Block instance when the instance is destroyed. That would be
a regression introduced a few commits back. That and the fact that File
objects created in Analyzer::Interface were not being freed combined to
represent a rather unhealthy 7k per run of Aesalon upon itself.

Now, the next step is to use the ScopeEvent framework set up. A class that
tracks the sent Scopes must be created; probably called ScopeManager. This
class will interact closely with the ptrace portal for parsing the memory maps
files for the running process, as well as the Analyzer framework for parsing
files in said maps. However, I am uncertain where it should be placed . . .