Attempted to implement a system of removing duplicate CanvasObjects.
Unfortunately, there are a few issues . . . for example, every so often, an
object that has its next and prev pointers set to itself appears . . . as one
can imagine, that is not healthy . . .

I think that the Canvas system needs an overhaul. One other thing that
definitely needs to happen: the main canvas should always be updated in real-
time. The "Real-Time updating" option should be to attach the viewport to the
newly-updated data.