Changed Canvas (and indeed, much of the GUI) around dramatically.
In essence, the results of these changes are: the local canvas is always
representative of the state of the memory of the monitored program. Events are
handled directly by the Renderer instances, meaning that the canvas is always
updated in real-time.

However, something that still needs to be done is to render the updated
regions of the Canvas into RenderedCanvas objects so the updates can be
properly displayed . . .