Finished most of a partial-updating system for painting Canvas objects.
In essence, when a Renderer notices an event, it marks a region of the Canvas
as needing an update. Every n milliseconds (configurable, of course, in the
future, but currently 500), the Viewport looks through this list, and requests
that the updated regions be painted. Nothing too fancy, but very useful.