I feel rather silly to not have noticed valloc() and associated functions.
I mean, there are even references to them in the malloc(3) man page . . . and
yet, I managed to avoid noticing them . . .

So, there is something else to add to the list. The first item on the list,
however, needs urgent attention: for some reason, the GUI is barely processing
600 events per second on the fibonacci benchmark. In other words, it is taking
about four seconds to execute the program, then three minutes to process the
resulting events . . .

I am not certain what is causing this rather large slowdown. It only started
when I upgraded my version of Qt from 4.5.3 to 4.6.2, but I have looked
through the changelog for Qt to try to find what could be causing this.
Nothing obvious is present . . .