Finally, finished a decent implementation of a scope resolution system.
Of course, it is still very slow (I think there is a linear operation in there
somewhere that I will have to fix . . .), but it works . . . but C++ names are
not demangled properly.

Time to read up on demangling, methinks. I cannot really use nm, since nm
appears to not read the same executables that my own ELF parser does. Not
certain why, but I can prove it empirically . . .