Implemented most of a backtrace transfer system.
However, there appears to be a rather large issue with the deserialization. As
per usual, the second deserialization computes a backtrace with a random
number of elements. I have absolutely no idea of why this is happening -- an
analysis of the transmitted data shows no oddities, but the deserialization is
insisting there are . . .