Changed around the overload library a little.
Timestamps are now gathered within the overload library, rather than the
monitor proper. As a result, there is a little more of a speed impact
introduced, but the timestamps are much more accurate.

As a side note, the backtrace system is far too inefficient at the moment. It
needs serious reworking, methinks. One important step is improving the speed
the backtrace gathering, but that is not exactly top priority at the moment. I
suppose I will have to try to profile the build once or twice, then see what
happens . . .

Do note that the GUI currently does not display any scope information, but it
does receive and process it.