Improved backtrace support slightly.
I believe I am almost ready to tag a commit as version 0.3.0 and release it
again over the internet. However, I would like to optimize the backtrace
system slightly first. To quote from the gprof output:
  %   cumulative   self              self     total
 time   seconds   seconds    calls   s/call   s/call  name
 29.51     12.61    12.61 660738993     0.00     0.00
    Analyzer::StorageManager::get_from_list(long, char const*) const
 24.23     22.96    10.35   106742     0.00     0.00
    Analyzer::File::get_symbol_for(unsigned long)
 16.62     30.06     7.10 2643001128     0.00     0.00
    Analyzer::StorageManager::dereference_attribute(long) const
 12.86     35.55     5.50 660738975     0.00     0.00
    Analyzer::StorageManager::get_child(long, char const*) const
  7.19     38.62     3.07 991119308     0.00     0.00
    Analyzer::StorageManager::dereference_string(long) const
  7.02     41.62     3.00       10     0.30     0.30
    Analyzer::StorageManager::new_string(long)
  2.01     42.48     0.86     9683     0.00     0.00
    Analyzer::StorageManager::create_child(long, char const*)

According to gprof, Aesalon is spending 99.4399% of its processing power
within the Analyzer module. dereference_attribute() is apparently called a
grand total of 2,643,001,128 times during the course of the monitoring of a
fairly simple program.

Some serious optimization is needed . . .