I am considering a change to the GUI storage tree . . .
Basically, rather than each leaf node storing the blocks that originate in
that leaf, instead each leaf stores a pointer to the blocks that pass through
the leaf. This does not work very well for small nodes (e.g. the 256-byte size
that it currently is), but would work for pages (e.g. 4K blocks). Assuming
that the maximum size of a block is about 1GB (probably reasonable), then no
more than about 8MB of memory would be used to represent this block. And that
is with a very, very large block of memory. Typical blocks would probably take
less than 16KB to represent . . .

There's a tradeoff here, regarding tree depth. The smaller the tree depth, the
less memory used to represent each block. The larger, the less memory is used
when a new iteration of the tree is created . . .