Stress-tested Aesalon on the new random_test.
I monitored the random test with Aesalon for approximately five hours, and it
seemed to work very well. At about 200 minutes in, the memory usage was about
30MB; perfectly acceptable . . .

However, the bug I thought I'd fixed is popping up again . . . for some
reason, data is not being pushed back onto the queue in the NetworkReceiver.
I will have to look into that.