Began implementing a basic visualization filter system.
In essence, when completed, all data will be passed through a filter system
before being visualized. What this means is that it is possible to select
only a region of the program to be visualized (e.g. all backtraces matching
the regex "^ll_", if you have a linked-list implementation that has
functions beginning with ll_. That would allow you to track only the memory
allocated by your linked-list implementation. Rather useful; do you not agree?