A slight change to the current shm system is required.
Basically, the entire shm cannot be used to store data, as two semaphores --
and perhaps other data as well -- have to be stored as well. As such, an
offset must be stored as the first element in the shm, as a 32-bit integer,
stating the position that the transferrable data actually begins. Semaphores
will be set on the begin/end data offset pointers, as well. It is assumed that
the overload library and the monitor will not be reading and writing to the
same memory at the same time. I am not sure how safe that assumption is, but,
well, we shall see.