The SysV IPC interface will not work very well, unfortunately.
The main issue is the limited buffer size. On Linux, without super-user
privileges, it is impossible to set a buffer of more than 8192 elements; which
is not much better than the 1500 that the UNIX pipe buffer could hold.

The circular buffer is still the best bet, I think. Calling sched_yield() will
work to allow the overload library or the overload parser to create or receive
more data, whichever the situation calls for. Coupled with a nanosleep() for
1000 nanoseconds (1 millisecond), the CPU usage should be minimal. Still not
optimal, but it should work . . .