Re-write it is.
What I am thinking about monitoring, then. The most obvious is memory usage,
regarding which I will replicate the current system. Second, interprocess and
network I/O, which will be easy to do via read/write and assorted functions.
CPU usage, which will be sampled every few hundred milliseconds based on how
much CPU time the process has taken up. Shared memory maps -- address and
sizes, not accesses . . . Thread creation, child process creation, and perhaps
amount of mutexes/semaphores. The last would be rather difficult, but it
should be possible. Not sure how useful it'd be, though . . .

Now, all of this will be accomplished by a series of overload libraries, not
just the single one that Aesalon currently uses. There will be a "controller",
or a single library that handles I/O to the listening monitor, and everything
else will be handled by separate libraries. This way, only libraries can be
loaded at start-up, reducing speed impact or to reduce the needless gathering
of information. This should hopefully also create an additional layer of
flexibility.

Of course, I'll still have to think about how that library will actually
communicate with the other libraries, but it should definitely be possible to
provide a single function to do so.