A basic design for the new layout . . .
In essence, I'd like to do everything using dynamic libraries. I realize that
doing it like this will slow things down slightly, but I feel that the trade-
off is worth it.

Monitoring modules will be installed. Each module has three responsibilities.
First, it must gather the data using an overload library. Second, it must
process the data in the monitor. Third, it must visualize the data in the GUI.
This triad of responsibilities means that the monitor and GUI are naught but
frameworks: the "real" work is done in the modules. Which is where the work
belongs.

I need to think up better names for the three sections, as well, while I'm at
it. "Overload library", "monitor", and "GUI" are all a bit too generic.

Such a set-up also makes it much easier to maintain the various sections of
the source tree. It does mean, though, that I have to be very careful about
changing the API. So, let me lay down some basic guidelines for each API . . .

-- Overload library (C API):

Data structures:

typedef struct {
    uint32 dataSize;
    void *data;
} DataPacket;
    Used to send data packets to the monitor library.

Macros:

AESALON_MODULE_ID(x): used once to denote the module ID. Must be unique.

Fuctions:
int AesalonSendData(DataPacket *packet): Sends data to the monitor via a
    shm. Will block if shm is full. If data->dataSize is larger than the shm
    size, the data will not be sent, and -1 will be returned. 0 indicates
    success.
int AesalonShouldGather(void): Returns 1 if main() has been called in the
    monitored program, 0 if it has not. Use to ignore pre-execution data.
uint64 AesalonGetTimestamp(): Returns a timestamp, accurate to the nanosecond.

-- Monitor library (C API):

DataPacket structure copied from overload library API.

The following functions must be exported to provide a valid library:

- AesalonInitModule(): Used to intialize the module.
- DataPacket *AesalonModuleDataReceived(DataPacket *): Called when a data
    packet is received. Should return a static DataPacket instance that does
    not need to be freed, with dynamically-allocated data.
- AesalonQuitModule(): Called when data gathering stops.

I'll leave the GUI API for later. There is an issue with the monitor API, and
that is the fact that both sections need to be able to interact with each
other. The monitor needs to be able to call AesalonModuleDataRecieved, and the
module should be able to call other functions like "AesalonGetScopeForAddress"
when required. I'll have to look up how I can do that, but I should be able to
provide a nice, clean, C API that can be used for a number of things.