Started buildsystem and framework for the module interface.
I still need to decide on terminology. I think I'll keep "monitor" the same,
and I'll probably use "visualizer" for the GUI, but I need to think of what to
call the loadable module layer. Of course, there are actually three layers,
but I need a name for the lowest layer.