Finished basic buildsystem for the modules and the monitor.
The terminology that I have decided on is to name the three layers the
collection layer, the monitoring layer, and the visualization layer. This
translates to the module's Collector library, the Monitor library, and the
Visualization library.

The Monitor and Visualization libraries will be C++ APIs, not C APIs. That is
simply because I feel that they will both be complex enough to benefit from
such a setup. I'll have to set down, in concrete, exactly how that API will be
accomplished, but I think I will simply set up a singleton-style instance that
the module libraries can access on demand.