Filled in a few blank spots in the framework.
I've been thinking about the configuration system, and I've come to the
conclusion that, this time around, I want to have a centralized configuration
system, multi-tiered of course. The first layer is the global configuration,
second is the per-user configuration, third is the per-directory
configuration, and fourth is the command-line options.

Having this centralized configuration system means that it is very easy to
make sections configurable; all that is required is a quick addition to the
default configuration list and a few function calls. Quite easy, really . . .