A design that I think will work, that is not too complex . . .
It involves several semaphores, unfortunately. In essence, there is one
central controlling semaphore, called the data semaphore. This is incremented
when a packet is placed in the ring buffer, and decremented when a packet is
removed. Another semaphore is used as a mutex to ensure that only one thread
is accessing the data-start pointer, to make sure that only one thread is
sending data at a time. Another is used for the data-end pointer.

The tricky bit is when the collection library is waiting on the monitor to
empty the buffer enough for a packet to be sent. This is solved by usage of
yet another semaphore: the overflow semaphore. The overflow semaphore is set
if a thread is waiting for more space to become available, and is cleared
every time a packet is received in the monitor. The waiting thread can set the
semaphore again if even more space is required. (Sending a 10KB packet when
the packets being received are only 64 bytes is an example of this behaviour.)

Overall, four semaphores are used for communication. It could definitely be
worse.

I checked the maximum value of a semaphore on three Linux computers. On both
64-bit systems and on the 32-bit system, the value was INT_MAX. According to
the webpage http://www.sbin.org/doc/glibc/libc_34.html, this will always be
the case in glibc. Those not using glibc, beware, I suppose . . .