Created a very basic Qt window for the visualizer.
I still have to think about the design of the visualizer, however. While the
module aspect is relatively obvious -- as is the network receiving bit -- the
GUI aspect is still unthreshed. I would like to avoid the system used in
previous versions; it is far too difficult to use. We'll see, I suppose.