The network framework in the monitor has been partially implemented.
Using netcat, it is now possible to view the dumped network data that is
produced by the monitor. It looks approximately correct, and I will have to
assume for the moment that it is.

The visualizer still needs quite a bit of work before it can be used for said
confirmation, though.