A few random changes to the visualizer.
Something I need to work out is a data path setup for the visualizer. I think
using a configuration file should work, but it seems too temporary. Qt can
compile resources directly into an executable, which should work for the time
being, but I can't get CMake to do it for me. Which is annoying.