Added a basic (and untested!) LogWrapper to the visualizer.
I think I have come up with a visualization output design that should satisfy
most the major requirements that have appeared . . .

A simple "visualize" function is provided by the VisualizerInterface class,
which is overloaded in the module type. This function is passed a
VisualizationPainter instance, which is a wrapper around QPainter and
something similar to the old CoordinateMapper class. Coordinates are given in
(time, data) axes, with time expressed as a uint64 and data as a qreal, just
as before.

The painting system will be set up relatively similar to before: the visualize
function will paint only what needs to be painted. The module interface will
also provide a function similar to markNeedUpdate(), to mark a section of the
visualization as needing an update, probably due to new data being received
from the monitor. This visualization will be done in a separate thread -- one
visualization thread per visualization, actually, for maximum performance --
meaning that modules will have to be thread-safe. This is a detail that I
would definitely prefer to avoid, but . . . really, I don't have a choice. It
is too much to ask to have all the visualization performed within a single
thread. Especially if there are three or four visualizations displayed
concurrently.