There is a slight problem with the current visualization design.
The issue is regarding the lack of a layer between the Visualization object
and the visualization modules. This lack means that files including the
visualization module interface wind up also including some of the Qt headers,
which is quite obviously suboptimal. Therefore, a simple VisualizationWrapper
interface needs to be implemented.