Very basic CPU usage visualizations can now be produced by the visualizer.
Nothing very special -- just connected line graphs -- but it is a start. One
feature that I'd like to implement is built-in support for line graphs in the
VisualizationWrapper or someplace similar. This would make the job of the
modules that much easier, helping to reduce the redunancy of each module.