Began implementing a basic Controller and distributed rendering setup.
However, I'd like to change things around a little. I would like to have a
QThreadPool managing the rendering; this way, multiple visualizations can be
rendered concurrently, and even multiple sections of the same visualization.

This might speed things up a little, and it also reduces the number of idle
threads sitting around. However, the next challenge . . .

Real-time updating. This will be handled in a similar manner to what the
previous iteration did. Every n milliseconds, a timer will fire, causing the
viewport to shift and a rendering request to be generated. Problem is: where
should this timer be located? Should it go in the VisualizationController? The
Widget? The Visualization itself?

The best bet seems to be the Controller -- that is, assuming that I create one
Controller object per Visualization. I could potentially create only one
Controller, at least how its design is currently looking.

But one Controller per Visualization is not unrealistic, I suppose. If that is
the case, then the Widget will have to have a reference to the Controller,
since the Widget is where the update time will be set . . .