Mostly implemented multi-threaded rendering.
There are three issues, however. The first is a weird occurance:
subVisualization() is attempting to create visualizations with a negative y-
size. No clue why that is happening . . . but it is.

The second is a Qt weirdness. QPixmaps, due to their design, cannot be used
outside of the main rendering thread. This is very, very annoying . . . but
the end result is that Visualizations now use QImages instead. Unfortunately,
QImages are stored in the client memory, not the X server memory, making them
slower to blit . . . much slower.

The third issue: for some reason, the mutex isn't controlling access to the
QImage properly. Sometimes the lock() function is called twice, creating two
separate QPainters -- actually, an invalid one overwriting the perfectly valid
one -- and in general causing havoc. Not the good kind of havoc, either . . .