Well, it does help if the painting rect is normalized.
And that concludes it: distributed visualizations are now implemented. This
framework should, in theory, allow for a much more extensible setup, and also
allow the utilization of more than just a single core per visualization. It
will also allow a visualization to be split into sections and rendered
individually. What fun that could be . . .