There is an issue, perhaps an unrelated one.
Too many conversions between data coordinates and screen coordinates are
performed. The result is several Visualizations that have a render-size of
zero on one of the axes. Very annoying . . . but, with the current design,
there is very little I can do except enforce a minimum size of 1 . . . which
is inherently a very, very bad idea.