Many changes, but the gist is that zooming and moving now partially work.
"Partially" means that the panning doesn't automatically generate the small
exposed painting requests. I think it's something to do with the double
conversions, as I mentioned before.

However, with zooming, it is possible to create a sort of repainting . . . but
it is definitely suboptimal. Also: I need a way to abort the rendering of
intermediate zooming steps. Generating twelve full visualizations for twelve
zoom steps is definitely not an optimal way to perform this sort of action.