Considered using getrusage() instead of clock_gettime().
The issue is that getrusage() returns the user CPU time in microseconds, not
in nanoseconds. As a result, the output it produces is less applicable, due to
the relatively small precision . . . but it is not subject (apparently) to the
sort of SMP madness that CLOCK_PROCESS_CPUTIME_ID is.

I think I will support both.