An idea that I had . . .
The concept of dependency modules -- modules that provide a set of functions
to be used by other modules. Consider, for example, a thread module, or an FD
module. The thread module could provide hooks to other modules to use
regarding (for example) thread creation, thread destruction, and attribute
changes. This would make the jobs of the other, thread-related modules much
easier, resulting in a smaller, more separated, even compact codebase.

So . . . the trick is how do set up these dependencies. The basic system is
simple: (if (A:B && !B) load(B)), for each dependency of each module. Where
should these module dependencies be specified, though?

I think it is time to vamp the module system. Rather than having all the
modules in a single directory, called "libcpuTimeCollector.so", etc, have a
configuration file in a directory named "cpuTime" that specifies the names and
paths of the collector, monitor, and visualizer modules. This configuration
file will also provide the dependency list.