Added the capability to append onto arguments and config file options.
Now, one can create a configuration file item such as "modules=cpuTime", then,
pass "--modules+=,threadMonitor" on the command line, appending threadMonitor
to the list of modules to be loaded. Of course, the support for appending
could potentially be improved via the usage of some sort of mechanism that
allows the detection of the separator character.