Implemented group support for command-line arguments.
I'm not particularly looking forward to implementing it in the conf-parsing,
mainly because it's going to be a copy-and-paste job. Which really annoys me.

Something I've noticed: I appear to be a little more verbose when I'm tired.
Not only in commit messages, but also in comments. Wonder why that is . . .