Modified Configuration to be much more compact.
Configuration's parsing code is now located in one place, rather than the
previous two instances -- one for the command-line, one for the configuration
files.

Now, the idea is that each module directory has two configuration files in its
root. One is something like "monitor.conf", the other similar to
"visualizer.conf". I'm not entirely certain if the visualizer configuration
file is required, as the only entry in the file will be the location of the
visualizer library . . . oh well.

Something else that would be a very good idea to implement is handshaking. The
ability to give Aesalon and the various modules version numbers to reduce
misunderstandings would be a nice feature to have around. Definitely something
to implement before the next release of Aesalon.