Added support for parsing module-specific configuration files.
There are only two catches with the system. One, I don't want to exposed the
collector and monitor filenames, but I don't want them to be internal either.
I have to decide about that, relatively quickly.

The second: if options are specified in the module configuration file, then
they will overwrite options given in the "regular" configuration files. The
order of precedence is something similar to global, user, local, module,
command-line. Which module configuration files cannot specify default values,
they have to be placed inside the other conf files.

Something I should implement is an include() function for the configuration
files. That would help make this a little easier.