Started to actually implement some of the new collector interface.
Nothing too special so far, but lookups by name are now supported, via a hash
table provided by hcreate_r() and friends. There are still plenty of
compilation errors still, but that should be alright for now.

One worry I have is that the collector is becoming heavily tied to GNU
extensions to the standard library. This places some limitations on the
portability of Aesalon, but I suppose it should not be too difficult to remove
many of these dependencies via a good set of #ifdefs.