Fixed the random segfaults that were occuring.
Now the module system is back to where it used to be. The visualizer is
capable of producing rudimentary visualizations, and everything seems to be
working relatively well.

Next step, I suppose, is improving the robustness of the visualizer's feature
set. The monitor is behaving well, and I don't think there is a lot more to
work on there at this moment.