Fixed it, finally.
When you're looking at something long enough to realize there's nothing wrong
with what you're looking at, then you realize the problem is somewhere else.

In this case, dataEnd was not being increased when the data was split into two
segments. The result is that the monitor/visualizer work perfectly now with a
4096-byte SHM.