Started re-implementing the configuration system.
The idea is to remove the many hacks that are currently in place and come out
with a nice, cleanly-designed configuration system. Integral to this design is
a tree-like setup for the configuration system and strings-only data storage.

Together with an "always-parse" setup (e.g. to always parse all of the module
configuration files), things should get quite a bit easier.