A thought regarding the layered binary tree . . .
If every single event creates a new layer on the tree, then retreival and such
become very, very easy. Much easier. The issue is, of course, the increased
amount of memory involved.

However, when one charts it out . . . assume, for the moment, that the
monitored program has a total of one million allocations. Also assume that
each tree node is 64 bytes (it is, in reality, more like 40 bytes), and that
the tree depth is also 64. Then, for each event, 64 * 64 (4096 bytes, or 4KB)
is consumed. Therefore, 4GB of storage is required.

This sounds like a very large amount of memory, and that is because it is. The
key here will be path compression. Using path compression, the average tree
depth can be assumed to be less than 20 (20: 1048576 unique blocks in memory
at any one given time, which is overkill), then the storage required drops to
1280 bytes per event. For a smaller setup (100,000 blocks, for instance), then
the maximum memory consumed at any one time is 103 MB, under the same set of
assumptions.

If more intelligence is added to the system, then it becomes actually rather
easy to drop the memory usage even further. One possible system is to create
new snapshots only when a block of memory has been allocated and released --
using a visitor-based system to detect differences in the tree -- one can drop
the memory usage under all but worst-case circumstances. The worst-case for
this is when a program is allocating and releasing blocks rapidly. Even if one
makes the assumption that ten blocks are allocated before one of the ten is
freed, then that is 90% gross space savings. Many programs have long periods
of allocations and releases (initialization and destruction, for example),
thus, for best-case there is only ~1KB extra memory used.

This is not accounting for the fact that currently, using the snapshot-per-5k-
events system, the events themselves must also be stored. Also, the system in
general should be much faster -- no tree reconstruction need take place to
find data. Instead, tree traversal occurs, which should be faster for all but
iteration over every single data element.

The logic required to implement this system is actually quite simple. When
modifying a block, simply check what its last-modified snapshot is. If it is
the same as the current snapshot, then create a new snapshot and use that. If
not, continue with the current snapshot.

Now to actually start to implement it.