A slightly different design . . .
The important bit here is that deletions (free/release events) no longer need
to be tracked. Each snapshot (TreeHead) instead stores a vector of pointers,
sorted by time of occurance, to the Block instance that was changed by that
event. Snapshots are then taken on a per-event basis (per n events), and then
accessing the tree is relatively easy.

Should be interesting.