Something of a problem.
Besides the obvious issue mentioned in the last commit, a problem has arisen
regarding the default data range of a visualization. Indeed, this is a
fundamental issue that was never properly resolved in the previous iteration.

How should the visualizations operate? Should they be free viewports to an
open visualization? Should they be limited according to the data that has been
received and processed? Should they be automatically controlled?

It's a question that needs more than a few seconds of thought.