Started to implement a heartbeat-based system for automatic updates.
This automatic update system would allow Visualizations to limit their
viewports to ranges of already-present data. There is an issue, however.
Rather, two . . .

First; adding the heartbeat system as I am now is as a hack, not as a properly
designed system. As such, it will probably complicate matters to a point where
it may not be worth the trouble.

Second, to implement it properly would rely upon a substantial re-write of
much of the current codebase. I would have to tie visualizations to specific
renderers, and also tie renderers to the data caches; which is hardly an
optimal system.

Thus, the issue still stands. To clarify; the exact definition of the problem
is to provide a system that allows the shifting of visualizations to display
the most recent data (if in real-time mode) and also to create a system that
allows for this to be done in an efficient manner . . .

The best setup that I have been able to construct in my musings to this point
is, in essence, a heartbeat system. A thread inside the collector interface
(or even the monitor) sends off events every n milliseconds (I thought a good
value would be 10 or perhaps 50), which is processed by the visualizer and
renderers, updating the range of the visualizations. However, the tricky bit
is actually updating the range.

The heart of the issue, so to speak, is the current rendering system. Each
renderer is given a temporary -- temporary, note -- visualization on which to
paint data. This works well for non-realtime updating, but fails in real-time
updating, where the renderer creates new data as it is received . . . the
problem is the current one-renderer-per-visualization-type system; replacing
it with a one-renderer-per-visualization style system would probably work more
effectively.

However.

The factory system currently in place would not exactly take very well to that
proposal; in fact, a complete re-design would probably have to take place.

Thus, one redesign requires another. The question is, what should I redesign?