One possible solution.
If the VControllers do their jobs properly, then it should be possible to
implement a dual-heartbeat system, each ticking off two different things.

First would be the current heartbeat, updating the visualizer modules to the
most recent timestamp in the monitored program. Second would be the update
heartbeat, the timer that fires off every t milliseconds to tell the
VControllers to update their visualizations. This second heartbeat would be
session-specific (perhaps even SessionDisplay-specific?), and would have to be
set to a relatively high value, say 100ms. This would trigger the VController
to tell the associated Renderer to find the most recent heartbeat (stored in
the AbstractDataCache class) and to paint the newly updated region.

In addition, the "TotalRange" member of Visualization will be moved into
VController. DataRange will still be relatively interesting to calculate.

Now, on an unrelated note. The naming scheme used in the Renderer section of
the visualizer has a non-constant naming scheme. This is not really a good
thing . . .

So. The end result . . . I need to rename a class or three. The question is,
should they be named RendererAbstractFactory, or AbstractRendererFactory?