Changed most of the uses of uint64_t for timestamps into Timestamps.
Me, being rather shortsighted, never stopped to think about the possibility of
using relative timestamps anywhere. As a result, I apparently never questioned
my use of uint64_t as a data type for storing timestamps . . .

Now that I am using relative timestamps in a few places . . . well. Timestamp
(AC_Timestamp in the collector, of course) is a typedef for int64_t. The range
of an int64_t storing nanoseconds since the UNIX epoch is approximately 292
years; which should be good enough for now.